# docs/openapi/stigmem.json
{
  "components": {
    "schemas": {
...
      "GardenCreateRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "quarantine": {
            "default": false,
            "description": "Create as a quarantine garden (Spec-08-Quarantine-Garden)",
            "title": "Quarantine",
            "type": "boolean"
          },
          "scope": {
            "default": "company",
            "title": "Scope",
            "type": "string"
          },
          "slug": {
            "maxLength": 63,
            "minLength": 1,
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name"
        ],
        "title": "GardenCreateRequest",
        "type": "object"
      },
...
      "GardenMemberUpdateRequest": {
        "properties": {
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "GardenMemberUpdateRequest",
        "type": "object"
      },
      "GardenRecord": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "garden_id": {
            "title": "Garden Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/GardenMemberRecord"
            },
            "title": "Members",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "quarantine": {
            "default": false,
            "title": "Quarantine",
            "type": "boolean"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "garden_id",
          "slug",
          "name",
          "scope",
          "created_by",
          "created_at"
        ],
        "title": "GardenRecord",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Reference node implementing the Stigmem v0.9.0a1 HTTP API \u2014 facts, federation, gardens, recall, subscriptions, audit, identity, content-addressed fact IDs. Cross-cutting features (tombstones, time-travel, multi-tenant) are opt-in plugins.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Stigmem Reference Node",
    "version": "0.9.0a1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/gardens": {
      "get": {
        "description": "List accessible gardens (Spec-02-Scopes-and-ACL).\n\nNode admins see all; others see only their gardens.",
        "operationId": "list_gardens_v1_gardens_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GardenRecord"
                  },
                  "title": "Response List Gardens V1 Gardens Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Gardens",
        "tags": [
          "gardens"
        ]
      },
      "post": {
        "description": "Create a new garden (Spec-02-Scopes-and-ACL). Creator is auto-added as admin.",
        "operationId": "create_garden_v1_gardens_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GardenCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GardenRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Garden",
        "tags": [
          "gardens"
        ]
      }
    },
    "/v1/gardens/{garden_slug_or_id}": {
      "delete": {
        "description": "Delete a garden (Spec-02-Scopes-and-ACL).\n\nRequires garden admin role. Facts are orphaned, not deleted.\nQuarantine gardens with pending facts cannot be deleted\n(Spec-08-Quarantine-Garden; HTTP 409).",
        "operationId": "delete_garden_v1_gardens__garden_slug_or_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "garden_slug_or_id",
            "required": true,
            "schema": {
              "title": "Garden Slug Or Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Garden",
        "tags": [
          "gardens"
        ]
      },
      "get": {
        "description": "Get a garden by slug or UUID (Spec-02-Scopes-and-ACL).",
        "operationId": "get_garden_v1_gardens__garden_slug_or_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "garden_slug_or_id",
            "required": true,
            "schema": {
              "title": "Garden Slug Or Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GardenRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Garden",
        "tags": [
          "gardens"
        ]
      }
    },
    "/v1/gardens/{garden_slug_or_id}/members": {
      "get": {
        "description": "List members of a garden (Spec-02-Scopes-and-ACL). Requires any member role.",
        "operationId": "list_members_v1_gardens__garden_slug_or_id__members_get",
        "parameters": [
          {
            "in": "path",
            "name": "garden_slug_or_id",
            "required": true,
            "schema": {
              "title": "Garden Slug Or Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GardenMemberRecord"
                  },
                  "title": "Response List Members V1 Gardens  Garden Slug Or Id  Members Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Members",
        "tags": [
          "gardens"
        ]
      },
      "post": {
        "description": "Add a member to a garden (Spec-02-Scopes-and-ACL). Requires admin role.",
        "operationId": "add_member_v1_gardens__garden_slug_or_id__members_post",
        "parameters": [
          {
            "in": "path",
            "name": "garden_slug_or_id",
            "required": true,
            "schema": {
              "title": "Garden Slug Or Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GardenMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GardenMemberRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Member",
        "tags": [
          "gardens"
        ]
      }
    },
    "/v1/gardens/{garden_slug_or_id}/members/{entity_uri}": {
      "delete": {
        "description": "Remove a member from a garden (Spec-02-Scopes-and-ACL). Cannot remove last admin.",
        "operationId": "remove_member_v1_gardens__garden_slug_or_id__members__entity_uri__delete",
        "parameters": [
          {
            "in": "path",
            "name": "garden_slug_or_id",
            "required": true,
            "schema": {
              "title": "Garden Slug Or Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entity_uri",
            "required": true,
            "schema": {
              "title": "Entity Uri",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Member",
        "tags": [
          "gardens"
        ]
      },
      "patch": {
        "description": "Change a member's role (Spec-02-Scopes-and-ACL). Requires admin.",
        "operationId": "update_member_role_v1_gardens__garden_slug_or_id__members__entity_uri__patch",
        "parameters": [
          {
            "in": "path",
            "name": "garden_slug_or_id",
            "required": true,
            "schema": {
              "title": "Garden Slug Or Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entity_uri",
            "required": true,
            "schema": {
              "title": "Entity Uri",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GardenMemberUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GardenMemberRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Member Role",
        "tags": [
          "gardens"
        ]
      }
    },
    "/v1/gardens/{garden_slug_or_id}/promote": {
      "post": {
        "description": "Promote a quarantined fact to a target garden (Spec-08-Quarantine-Garden).\n\nRequires quarantine:moderator or admin role.",
        "operationId": "promote_fact_v1_gardens__garden_slug_or_id__promote_post",
        "parameters": [
          {
            "in": "path",
            "name": "garden_slug_or_id",
            "required": true,
            "schema": {
              "title": "Garden Slug Or Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuarantinePromoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Promote Fact V1 Gardens  Garden Slug Or Id  Promote Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Promote Fact",
        "tags": [
          "gardens"
        ]
      }
    },
...
  }
}